Create server.py
The server file, which you need to run in order for the clients to connect to each other! By default the IP is set to localhost (127.0.0.1) and the port is set to the default TCP port (9090). If you want to use this publicly then you will need to change these values, which you can do yourself or find guides on how to do online!