Create client.py
The ClientGUI app that allows the client to send messages, and also has a nice GUI! Like the server script by default the IP is set to localhost (127.0.0.1) and the port is set to the default TCP port (9090). The GUI was created by me using a guide from YouTube, so I don't take any credit for that!